prevents block breaking
